refactor: remove Transformable and Actionable protocols from policy.py; add new protocols in protocols.py
